requirements.txt
google-genai
fastapi
uvicorn
pytest
pydantic
pandas
PyPDF2
tqdm
faiss-cpu
python-dotenv